Move transaction types to separate files.
