+++ crates/client/src/rpc/types/eth/transaction.rs
@@ -21,7 +21,9 @@
 use crate::rpc::types::Bytes;
 use cfx_types::{H160, H256, H512, U256, U64};
 use cfx_vm_types::{contract_address, CreateContractAddress};
-use primitives::{transaction::eip155_signature, Action, SignedTransaction};
+use primitives::{
+    transaction::eth_transaction::eip155_signature, Action, SignedTransaction,
+};
 use rlp::Encodable;
 use serde::Serialize;
