+++ crates/client/src/rpc/types/transaction.rs
@@ -9,7 +9,10 @@ use cfx_addr::Network;
 use cfx_types::{Space, H256, U256, U64};
 use cfxkey::Error;
 use primitives::{
-    transaction::Action, Eip155Transaction, NativeTransaction,
+    transaction::{
+        eth_transaction::Eip155Transaction,
+        native_transaction::NativeTransaction, Action,
+    },
     SignedTransaction, Transaction as PrimitiveTransaction, TransactionIndex,
     TransactionWithSignature, TransactionWithSignatureSerializePart,
 };
