+++ crates/primitives/src/lib.rs
@@ -50,10 +50,12 @@ pub use crate::{
     },
     storage_key::*,
     transaction::{
-        Action, Eip155Transaction, NativeTransaction, SignedTransaction,
-        Transaction, TransactionWithSignature,
+        Action, SignedTransaction, Transaction, TransactionWithSignature,
         TransactionWithSignatureSerializePart, TxPropagateId,
     },
     transaction_index::TransactionIndex,
     zero::Zero,
+};
+pub use transaction::{
+    eth_transaction::Eip155Transaction, native_transaction::NativeTransaction,
 };
